# Per device stream