--- a/cupy/cuda/stream.pyx
+++ b/cupy/cuda/stream.pyx
@@ -12,15 +12,25 @@ cdef object _thread_local = threading.local()
 
 
 cdef class _ThreadLocal:
+    # We keep both current_stream_ref and current_stream_stack because the
+    # former is also used when calling stream.use(). This bookkeeping enables
+    # correct rewinding when "with" blocks are mixed with ".use()" (though
+    # this is considered an anti-pattern). As the name suggested, stream
+    # lifetime is not tracked in current_stream_ref.
+
     cdef list current_stream_ref  # list of object
     cdef list current_device_id_stack  # list of int
-    cdef list prev_stream_stack  # list of list
+    cdef list current_stream_stack  # list of list
 
     def __init__(self):
         cdef int i, num_devices = runtime.getDeviceCount()
-        self.current_stream_ref = [None for i in range(num_devices)]
-        self.prev_stream_stack = [[] for i in range(num_devices)]
+        self.current_stream_ref = []
         self.current_device_id_stack = []
+        self.current_stream_stack = []
+        for i in range(num_devices):
+            default_stream = get_default_stream()
+            self.current_stream_ref.append(weakref.ref(default_stream))
+            self.current_stream_stack.append([default_stream])
 
     @staticmethod
     cdef _ThreadLocal get():
@@ -31,16 +41,18 @@ cdef class _ThreadLocal:
         return <_ThreadLocal>tls
 
     cdef void push_stream(self, stream, int device_id) except*:
-        prev_stream = self.get_current_stream(device_id)
-        self.prev_stream_stack[device_id].append(prev_stream)
+        assert device_id >= 0
+        self.current_stream_stack[device_id].append(stream)
         # record device_id to prevent from popping the wrong stream at exit
         self.current_device_id_stack.append(device_id)
         self.set_current_stream(stream)
 
     cdef void pop_stream(self) except*:
         cdef int device_id = self.current_device_id_stack.pop()
-        prev_stream = self.prev_stream_stack[device_id].pop()
+        self.current_stream_stack[device_id].pop()
+        prev_stream = self.current_stream_stack[device_id][-1]
         self.set_current_stream(prev_stream)
+        assert len(self.current_stream_stack[device_id]) >= 1
 
     cdef set_current_stream(self, stream):
         cdef intptr_t ptr = <intptr_t>stream.ptr
@@ -54,9 +66,6 @@ cdef class _ThreadLocal:
         if device_id == -1:
             device_id = runtime.getDevice()
         stream_ref = self.current_stream_ref[device_id]
-        if stream_ref is None:
-            stream_ref = weakref.ref(get_default_stream())
-            self.current_stream_ref[device_id] = stream_ref
         return stream_ref()
 
     cdef intptr_t get_current_stream_ptr(self):