… cupy_backends/cuda/stream.pxd
… cupy_backends/cuda/stream.pyx
… docs/source/user_guide/basic.rst
… tests/cupy_tests/cuda_tests/test_stream.py