Close #3980. Fix #5143. Fix #5145.

For the reason in https://github.com/cupy/cupy/issues/3980#issuecomment-831487066, we call `getDevice()` every time we need to look up the device.

***UPDATE:*** major changes:
- Stream references are kept when entering the `with stream` context to avoid dead references, but not when calling `stream.use()` (which still keeps only the stream weakref)
- The stream stack now also records the *current* stream so that if a stream is deleted it can fall back correctly.

TODO:
- [x] Add tests