@leofang 2021-05-05
I feel the current implementation (up to 105503a) adds too many calls to `getDevice()` which should not be necessary. Let me see what I can do.