@kmaehashi 2021-05-06
I feel TLS implementation is a bit complicated because we need to avoid creating `weakref`s everytime setting a stream. Maybe doing `self._weakref = weakref.ref(self)` in `Stream`/`ExternalStream` constructor may simplify the code?
(This is out of the scope of the original motivation of this PR, so you can ignore this comment 😄 )