@leofang 2021-05-06
> I feel TLS implementation is a bit complicated because we need to avoid creating `weakref`s everytime setting a stream.

@kmaehashi I don't think I get it, could you elaborate?

> (This is out of the scope of the original motivation of this PR, so you can ignore this comment 😄 )

That's OK, let me address it properly in this PR.