@kmaehashi 2021-05-06
@leofang 
> I feel TLS implementation is a bit complicated because we need to avoid creating weakrefs everytime setting a stream. > Maybe doing self._weakref = weakref.ref(self) in Stream/ExternalStream constructor may simplify the code?

What I meant is that if we create a weakref of the stream inside `__init__` of `Stream`/`ExternalStream` (i.e., let each stream instance hold the weakref to self), we can remove `get_current_stream_ref`/`set_current_stream_ref` APIs in TLS.