@leofang 2021-05-09
@kmaehashi Since I'm working on the stream stack, it'd be nice if the team could make a decision on #4337 🙂 Specifically, do we want to preserve the current behavior if the stream `s` is deleted after `s.used()` is called?