@leofang 2021-05-10
> it'd be nice if the team could make a decision on #4337 🙂 Specifically, do we want to preserve the current behavior if the stream `s` is deleted after `s.used()` is called?

The resolution is to keep the present behavior as is, so after `s` is deleted we fall back to the default stream.