@leofang 2021-05-12
@takagi @kmaehashi I need your input on another design decision. This time it's related to #4445 with the assumption that in the near future the two stream modules `cupy/cuda/stream.pyx` and `cupy_backends/cuda/stream.pyx` remain separate.

Currently we store the current stream pointer in the thread-local storages of both modules. I would like to reduce the repetition. Should we separate the roles of the two storages in the following fashion?
- In `cupy/cuda/stream.pyx` we store the stream stack and weak refs (see also https://github.com/cupy/cupy/pull/5172#discussion_r627048001)
- In `cupy_backends/cuda/stream.pyx` we store the current stream pointer

The only change IIUC is to remove `self.current_stream` from the the thread-local storage in `cupy/cuda/stream.pyx`. I hope I am not missing anything...