@leofang 2021-05-14
> The only change IIUC is to remove `self.current_stream` from the the thread-local storage in `cupy/cuda/stream.pyx`. I hope I am not missing anything...

I made the judgement call to do this in 9cdd026.