@kmaehashi 2021-05-18
Only the bug-fix part (#5156) should be backported.