@leofang 2021-05-18
I think most concerns are addressed by now, if not please let me know.

I do want to point out the performance impact of this PR. Below is a simple script to time 1. synthetic workload with streams, 2. `with stream` context manager, and 3. `stream.use()` switch:
```python
import cupy as cp
from cupyx.time import repeat


N = 32
runs = 128
streams = []
for i in range(N):
    streams.append(cp.cuda.Stream())

def stream_test1():
    result = []
    for i in range(runs):
        with streams[i % N] as s:
            a = cp.random.random((2**20,))
            b = cp.sum(a)
            result.append(b)
    for i in range(N):
        streams[i].synchronize()

def stream_test2():
    cp.cuda.Stream.null.use()
    for i in range(runs):
        with streams[i % N] as s:
            pass
    cp.cuda.Stream.null.use()

def stream_test3():
    cp.cuda.Stream.null.use()
    for i in range(runs):
        streams[i % N].use()
    cp.cuda.Stream.null.use()

print(repeat(stream_test1, n_repeat=100))
print(repeat(stream_test2, n_repeat=1000))
print(repeat(stream_test3, n_repeat=1000))
```
Before this PR (current master):
```
stream_test1        :    CPU: 9198.319 us   +/-25.499 (min: 9153.547 / max: 9265.513) us     GPU-0: 9203.953 us   +/-25.540 (min: 9158.656 / max: 9271.200) us
stream_test2        :    CPU:   67.898 us   +/- 1.714 (min:   66.936 / max:   85.735) us     GPU-0:   70.155 us   +/- 1.727 (min:   68.128 / max:   88.064) us
stream_test3        :    CPU:   33.462 us   +/- 0.653 (min:   33.110 / max:   45.828) us     GPU-0:   35.764 us   +/- 0.887 (min:   34.816 / max:   48.992) us
```
With this PR:
```
stream_test1        :    CPU: 9203.129 us   +/-28.180 (min: 9139.099 / max: 9281.473) us     GPU-0: 9208.702 us   +/-28.213 (min: 9145.504 / max: 9287.552) us
stream_test2        :    CPU:  103.754 us   +/- 1.722 (min:  102.756 / max:  124.242) us     GPU-0:  106.116 us   +/- 1.818 (min:  104.448 / max:  125.856) us
stream_test3        :    CPU:   48.798 us   +/- 1.327 (min:   44.627 / max:   64.436) us     GPU-0:   51.176 us   +/- 1.393 (min:   47.008 / max:   66.624) us
```
The performance impact is negligible with sufficient workload (`stream_test1`). However, both `stream_test2` and `stream_test3` show an additional ~50% overhead, though the overhead increase *per stream* is on the order of 1 us.