@kmaehashi 2021-05-21
I noticed that the current stream is lost when pushing a stream.

```py
>>> s1 = cupy.cuda.Stream()
>>> s2 = cupy.cuda.Stream()
>>> s2.use()
<Stream 94167172638448 (device 0)>
>>> with s1:
...     pass
... 
>>> 
>>> cupy.cuda.get_current_stream()
<Stream 0 (device -1)>  # <--- should be  s2
```