@takagi 2021-05-21
To make it work well with `use()`, I think, we need to shift the values to push into `current_stream_stack`. That is, we have to push the old current stream here instead of the new `stream`:

https://github.com/cupy/cupy/blob/8d163896d9e7fa39eea6b6ab24b4c7b6e0d97bd6/cupy/cuda/stream.pyx#L45

as well as changing `current_stream_stack`'s initial value for each device to `[]`.

https://github.com/cupy/cupy/blob/8d163896d9e7fa39eea6b6ab24b4c7b6e0d97bd6/cupy/cuda/stream.pyx#L33