@leofang 2021-05-21
@kmaehashi @takagi Maybe I still misunderstood you. I thought by closing #4337 we agreed to not push a stream to the stack if it's used via calling `stream.use()`? With this PR, [your snippet](https://github.com/cupy/cupy/pull/5172#issuecomment-845764323) is expected.