@takagi 2021-05-21
I understand #4337 that `with stream:` remembers the (previous) current stream when it has been called regardless it was set by `use()` or another stream context. On the other hand, `stream.use()` doesn't remember it so doesn't push it on the stack.