@leofang 2021-05-22
If we push the "previous" current stream to the stack (the old approach), we lose the ability to rewind to the last stream in use when the stream made in use via `.use()` is deleted (#5143). I think it's fundamentally incompatible unless we push the stream to the stack unconditionally (#4337).