@takagi 2021-05-22
Thanks, now I got #5143. I needed to work by my hands to understand these related PRs... An extra slot for each stack level that shadows the stream context when `use()` might work, but I'm not confident without writing codes for myself.